properly merged PRs and combined login functionality with general dashboard & routing
