+++ src/App.js
@@ -5,6 +5,7 @@ import NavBar from './components/NavBar';
 import GeneralOwnerDashboard from './screens/general/GeneralOwnerDashboard';
 import Onboarding from './screens/onboarding/Onboarding';
 import Login from './screens/auth/Login';
+import SignUp from './screens/auth/SignUp';
 
 function App() {
   return (
@@ -14,6 +15,7 @@ function App() {
 
         <Switch>
           <Route exact path="/" component={Login} />
+          <Route path="/signup" component={SignUp} />
           <Route path="/onboarding" component={Onboarding} />
           <Route path="/dashboard" component={GeneralOwnerDashboard} />
           <Route>Not Found - 404</Route>
