--- src/RoutingApp.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import './styles/index.css';
-import Main from './loginComponents/Main';
-
-function RoutingApp() {
-  return (
-    <div>
-      <Main />
-    </div>
-  );
-}
-
-export default RoutingApp;
